إصلاح: تواريخ بصيغة رقمية فقط مع أرقام عربية للغة العربية
- عربي: ٤/١٢/٢٠٢٥ (أرقام عربية)
- إنجليزي: 12/4/2025 (أرقام غربية)
- جميع التواريخ الآن بصيغة رقمية موحدة بدون أسماء الأشهر
- التطبيق على جميع صفحات الموقع

+++ app/dashboard/attendance/page.tsx
@@ -321,7 +321,7 @@ export default function MyAttendancePage() {
                       </div>
                       <div className="flex items-center gap-2 text-gray-300">
                         <span>📝</span>
-                        <span>{formatDate(new Date(attendance.registeredAt), language, 'short')}</span>
+                        <span>{formatDate(new Date(attendance.registeredAt), language, 'long')}</span>
                       </div>
                     </div>
